requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
Jinja2==3.1.4
SQLAlchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.1
alembic==1.13.2
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
APScheduler==3.10.4
python-dotenv==1.0.1
itsdangerous==2.2.0
httpx==0.27.0
Pillow==10.3.0
